Don't put the original article in the list
This change removes the originally posted article from the list of
sources (one of the most frequently mentioned issues). One of the
benefits of the duplication was that it would prove if the post's
title differed from the article's, so the change explicitly mentions if
the title has been changed at the beginning of the comment.